Update components managing
